// src/app/home/models/tb-product.ts
export class TbProduct {
  prodSerial?: number= null;
  forbom?: number= null;
  partFlag?: number= null;
  partImpSeq?: number= null;
  partImpSerial?: number= null;
  partNameEng?: string= null;
  partNameTh?: string= null;
  partNo?: string= null;
  partUnit?: number= null;
  partWeight?: number= null;
  partuseDesc1?: string= null;
  partuseDesc2?: string= null;
  partuseDesc3?: string= null;
  partuseNo1?: string= null;
  partuseNo2?: string= null;
  partuseType?: number= null;
  pplStaticCode?: string= null;
  pplTariffSeq?: string= null;
  pplTariffDesc?: string= null;
  
  pplTariffNo?: string= null;
  prodFlag?: number= null;
  showFlag?: number= null;
  supSerial?: number= null;
  weightUnit?: number= null;
  
  offset?: number= null;
  limit?: number= null;  
  createBy?: string= null;
  updateBy?: string= null;


}